update home 23 and cart 1

+++ main_app/models.py
@@ -113,9 +113,9 @@ class Product:
                 "singles_stock", "carton_bag_stock", "description", "slug", "is_divisible", "is_carton_bag_divisible")
     
     def __init__(self, brand_name: str, product_name: str, size: str, product_image: list, tags: list,
-                 retail_price: int, wholesale_price: int, is_discount: bool, discount_retail_price: int, has_bulk: bool,
+                 retail_price: float, wholesale_price: float, is_discount: bool, discount_retail_price: float, has_bulk: bool,
                  bulk_prices: dict, bulk_types: dict, nos_in_bulk: dict, bulk_images: dict, is_carton_bag: str,
-                 carton_bag_price: int, no_in_carton_bag: int, carton_bag_image: list, price_modified_date: datetime,
+                 carton_bag_price: float, no_in_carton_bag: int, carton_bag_image: list, price_modified_date: datetime,
                  singles_stock: int, carton_bag_stock: int, description: str, slug: str, is_divisible: bool,
                  is_carton_bag_divisible: bool) -> None:
         self.brand_name = brand_name
@@ -123,10 +123,10 @@ class Product:
         self.size = size
         self.product_image = product_image
         self.tags = tags
-        self.retail_price = retail_price
-        self.wholesale_price = wholesale_price
+        self.retail_price = float(retail_price)
+        self.wholesale_price = float(wholesale_price)
         self.is_discount = is_discount
-        self.discount_retail_price = discount_retail_price
+        self.discount_retail_price = float(discount_retail_price)
         self.is_divisible = is_divisible
         self.has_bulk = has_bulk
         self.bulk_prices = bulk_prices
@@ -134,7 +134,7 @@ class Product:
         self.nos_in_bulk = nos_in_bulk
         self.bulk_images = bulk_images
         self.is_carton_bag = is_carton_bag # Does Product come in Cartons or Bags?
-        self.carton_bag_price = carton_bag_price
+        self.carton_bag_price = float(carton_bag_price)
         self.no_in_carton_bag = no_in_carton_bag
         self.carton_bag_image = carton_bag_image
         self.price_modified_date = price_modified_date
@@ -197,14 +197,14 @@ class StaffCart:
     __slot__ = ("name_of_buyer", "staff_id", "items", "total_amount", "checkout_date",
                 "amount_paid", "created_date")
     
-    def __init__(self, name_of_buyer: str, staff_id: str, items: list, total_amount: int,
-                 amount_paid: int, created_date = datetime.now(), checkout_date = datetime.now()) -> None:
+    def __init__(self, name_of_buyer: str, staff_id: str, items: list, total_amount: float,
+                 amount_paid: float, created_date = datetime.now(), checkout_date = datetime.now()) -> None:
         self.name_of_buyer = name_of_buyer
         self.staff_id = staff_id
         self.items = items
-        self.total_amount = total_amount
+        self.total_amount = float(total_amount)
         self.checkout_date = checkout_date
-        self.amount_paid = amount_paid
+        self.amount_paid = float(amount_paid)
         # Note that amount owed can be negative, which will translate to giving customer change
         self.amount_owed = self.total_amount - self.amount_paid
         self.created_date = created_date
